+++ packages/agents-hosting/src/agent-client/index.ts
@@ -1,2 +1,2 @@
 export * from './agentClient'
-export * from './expressApi'
+export * from './agentResponseHandler'
